Refactor useOverflowDetection to single ref
